Move fork logic to separate file
