+++ adagucserverEC/CMakeLists.txt
@@ -245,7 +245,8 @@ add_library(
     CImageOperators/smoothRasterField.cpp
     CImageOperators/drawContour.h
     CImageOperators/drawContour.cpp
-    
+    fork_server.h
+    fork_server.cpp
 )
 
 target_include_directories(adagucserverEC PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} ${Cairo_INCLUDE_DIRS} ${FREETYPE_INCLUDE_DIRS} ${PostgreSQL_INCLUDE_DIRS} ${GDAL_INCLUDE_DIRS} ${PROJ_INCLUDE_DIR})
