+++ adagucserverEC/adagucserver.cpp
@@ -38,6 +38,7 @@
 #include "CDBFactory.h"
 #include "CConvertGeoJSON.h"
 #include "utils/serverutils.h"
+#include "fork_server.h"
 
 #include <sys/socket.h>
 #include <sys/un.h>
@@ -387,8 +388,6 @@ int _main(int argc, char **argv, char **) {
   return getStatusCode();
 }
 
-static const unsigned int nIncomingConnections = 5;
-
 int do_work(int argc, char **argv, char **envp) {
   /* Check if ADAGUC_LOGFILE is set */
   const char *ADAGUC_LOGFILE = getenv("ADAGUC_LOGFILE");
@@ -467,100 +466,13 @@ int do_work(int argc, char **argv, char **envp) {
   return status;
 }
 
-void handle_client(int client_socket, int argc, char **argv, char **envp) {
-  int recv_buf_len = 4096;
-  char recv_buf[recv_buf_len];
-  memset(recv_buf, 0, recv_buf_len * sizeof(char));
-
-  int data_recv = recv(client_socket, recv_buf, recv_buf_len, 0);
-  if (data_recv > 0) {
-    dup2(client_socket, STDOUT_FILENO);
-    setenv("QUERY_STRING", recv_buf, 1);
-
-    int status = do_work(argc, argv, envp);
-    // fprintf(stderr, "exiting, status=%d", status);
-
-    fflush(stdout);
-    fflush(stderr);
-
-    exit(status);
-  }
-}
-
-static std::map<pid_t, int> child_sockets;
-
-void on_child_exit(int child_signal) {
-  int child_status;
-  pid_t child_pid = wait(&child_status);
-
-  int child_sock = child_sockets[child_pid];
-  fprintf(stderr, "@ on_child_exit [%d] [%d] [%d] [%d]", child_pid, child_sock, child_signal, child_status);
-
-  write(child_sock, reinterpret_cast<const char *>(&child_status), sizeof(child_status));
-  close(child_sock);
-
-  child_sockets.erase(child_pid);
-}
-
-int run_server(int argc, char **argv, char **envp) {
-  int client_socket = 0;
-
-  signal(SIGCHLD, on_child_exit);
-
-  struct sockaddr_un local, remote;
-  int len = 0;
-
-  int listen_socket = socket(AF_UNIX, SOCK_STREAM, 0);
-  if (-1 == listen_socket) {
-    printf("Error on socket() call \n");
-    return 1;
-  }
-
-  local.sun_family = AF_UNIX;
-
-  CT::string socket_path(getenv("ADAGUC_PATH"));
-  socket_path.concat("/adaguc.socket");
-
-  strcpy(local.sun_path, socket_path.c_str());
-  unlink(local.sun_path);
-  len = strlen(local.sun_path) + sizeof(local.sun_family);
-
-  if (bind(listen_socket, (struct sockaddr *)&local, len) != 0) {
-    printf("Error on binding socket \n");
-    return 1;
-  }
-
-  if (listen(listen_socket, nIncomingConnections) != 0) {
-    printf("Error on listen call \n");
-  }
-
-  while (1) {
-    unsigned int sock_len = 0;
-
-    if ((client_socket = accept(listen_socket, (struct sockaddr *)&remote, &sock_len)) == -1) {
-      printf("Error on accept() call \n");
-      return 1;
-    }
-
-    pid_t pid = fork();
-    if (pid == 0) {
-      close(listen_socket);
-      handle_client(client_socket, argc, argv, envp);
-    } else {
-      child_sockets[pid] = client_socket;
-    }
-  }
-
-  return 0;
-}
-
 int main(int argc, char **argv, char **envp) {
   setvbuf(stdout, NULL, _IONBF, 0); // turn off buffering
   setvbuf(stderr, NULL, _IONBF, 0); // turn off buffering
 
   const char *ADAGUC_FORK = getenv("ADAGUC_FORK");
   if (ADAGUC_FORK != NULL) {
-    int server_status = run_server(argc, argv, envp);
+    return run_server(do_work, argc, argv, envp);
   } else {
     // normal flow without unix socket server/fork
     return do_work(argc, argv, envp);
